feat: persist webdav state
